[package] add mysql-server (#7384)
This is a patch to build the mysql server. So far I only have the
kirkwood target to test on, but it works great on that.

There seems to be some contention on whether Openwrt should have a
mysql-server package, but I don't see why it shouldn't at least have
the option. A full LAMP stack is now a reality.

Closes out https://dev.openwrt.org/ticket/7384

Again, I'm still pretty new to this, so any constructive criticism
would be appreciated.

Sorry for the attachment, but I wanted to make sure the whitespace and
lines got preserved.

Signed-off-by: Kyle Anderson <kyle@xkyle.com>